Refactor project page into separate route handlers
