+++ frontend/src/app/projects/[id]/scans/page.tsx
@@ -1,0 +1,37 @@
+"use client";
+
+import {
+  Card,
+  CardContent,
+  CardDescription,
+  CardHeader,
+  CardTitle,
+} from "@/components/ui/card";
+import { FileText } from "lucide-react";
+
+export default function ProjectScansPage({
+  params,
+}: {
+  params: Promise<{ id: string }>;
+}) {
+  return (
+    <div className="space-y-6">
+      <Card>
+        <CardHeader>
+          <CardTitle>All Scans</CardTitle>
+          <CardDescription>
+            Comprehensive list of all security scans for this project
+          </CardDescription>
+        </CardHeader>
+        <CardContent>
+          <div className="text-center py-8">
+            <FileText className="h-12 w-12 mx-auto mb-4 text-muted-foreground" />
+            <p className="text-muted-foreground">
+              Scans view coming soon
+            </p>
+          </div>
+        </CardContent>
+      </Card>
+    </div>
+  );
+}
